Update patch set 4
Patch Set 4:

Also, we can't always use the count of config errors from the tenant info endpoint because it's the total number of errors, whereas this page shows errors that are the result of server-side filtering.  So as soon as a user adds a filter, it's wrong.  We could use the total count only in the case where there are no filters, and that might still be marginally useful.  At any rate, it's complex and I don't think it's required for this change given the lack of infrastructure around this in other places.

Patch-set: 4