Don't let admin action dropdown move off screen
The admin action dropdown on the status page get displayed off the
screen if the toggle icon is at the very right end of it.
Avoid this by telling the Dropdown component its position.
A consequence of this is that the dropdown may overlap the change item
while being displayed.

Change-Id: Iaa7ae8586a2e2304ea94fcf3963e5e121c9dc10d