Remove nested templates from assert in tests
Ansible has fixed CVE-2023-5764 by further restricting evaluation
in "nested templates" (which I take to mean a jinja template string
inside of an ansible field that is already expected to be a template
string).  The field named "that" in the "assert" module is such a
field.

To address this, remove the nested template and just treat the
expression as a single jinja template.

Change-Id: I483cf2ec0f61e6484c2768acfb7ab8d9dd5c5117