Remove file matcher special case for empty file list
Some background:

We originally added special handling for the case where the files list
for a change had only the /COMMIT_MSG pseudo-file:

  Iad78d9eb8212beea3238728321c1ba74efa991e2

That adjust the all-files-matches-any skip-if handler (that's the
negative matcher, since renamed to irrelevant-files) to match that
case.  The theory is that in the case of a merge commit, we would get a
files list with only the commit message, and since we don't know what to
run or not to run, we should run everything.

Later we did some rework to the positive matcher (ie, "files" rather
than "irrelevant-files") to match refs with no file list:

  Icf5df145e4cd351ffd04b1e417e9f7ab8c5ccd12

I use refs here intentionally.  That change was intended to address the
issue of jobs not running for branch head queue items (ie, not
changes).  Follow its reference to
If7a3a7cc212c981529be086dadb8157f08bda342 to see the motivation.

Note that having no files list is different than having a files list
with only the commit message.  Yet for some reason that I can not
determine based on the written history in that change, the behavior was
extended to not only encompass the "no files list at all" case and also
includes the "files list with only the commit message" case.  This
changed the behavior of the files matcher to the current behavior where
if you say a job should only run on changes to "docs/" and you push up
an empty commit, it will run the job.

This does appear to be intentional in that change, but the reason was
not explained.  It may have been decided that was acceptable in order to
mimic the merge commit behavior of the negative matcher.

However, since that time, Zuul is now able to determine the files
affected by a merge commit (either from the code review system API, or
using its own mergers to perform the calculation).  We should no longer
get files lists for merge commits that have only the commit message
(assuming of course the underlying commits do actually change files).

What we *are* likely to see now is a new user behavior: Gerrit users who
use the Gerrit web UI to create a change will cause an empty change
(with only a commit message file) to be created as the first patchset.
Later files are added to new patchsets and behave normally, but that
initial patchset will cause all jobs in a repo to run.  This is both
inconvenient and counter to expectations (users who write files matchers
expect changes to trigger jobs only if they have those files).

While the above focuses on the "files" matcher, the same is also true
for the "irrelevant-files" matcher.  The use-case that required special
handling for empty commits as a proxy for merge commits is no longer
present there.

This change removes the special casing of empty commits in the file
matchers.

Change-Id: I75efae0f8820760dc956c39daa7a8dadcc97472e