Fix multi-scheduler timer unit test
This unit test gets the list of apscheduler jobs and inspects them,
but that list is returned in pending-job order, meaning that it may
be different based on whether the job has recently run.

To correct this, sort the jobs by the jitter, which we expect to be
consistent and different (these characteristics are asserted later
in the thest).

Change-Id: I052e88573121f88a016971fb6461e7ed8fff25a6