Avoid unecessary join with provides table
Under some circumstances, postgres may produce an inefficient
query plan for build queries.  For example, the build time
estimation query does not filter builds by "provides" values but
we join the table regardless because the method supports queries
that do.  In some cases, postgres might try to apply that join
early in the query plan, which can cause it to run very slowly.

To avoid this situation, let's not join the provides table unless
necessary for the query.  This should make the time estimation
queries more deterministic, as well as improving many other
potential build queries.

Change-Id: I2f85b92b4e6cb5a9c1cff13658692da170737a98