Add an index to speed build time estimation
This change adds an index to the build table that is intended to
speed the query used by the build time estimator.  This index
covers all of the fields it uses as well as the buildset id, so
that the dbms can use a single index to filter and join the build
table.

Change-Id: Ie4f0b285f7e0734417e77b762962e1897fbd1648