Gerrit driver: fix for topics containing white space
When using gerrit topcis containing white spaces, zuul fails to find the
changes contained in the topic because the query it builds does not
enclose the topic in quotes. So only the first word of the topic is
considered by the gerrit driver. Fixing this by quoting the topic in the
query.

Change-Id: I99d2890d317fb8424740e25d166d17381f1319c8