Replace python imp module with importlib
The imp module has been deprecated for quite some time. The importlib
module has been present since python 3.6 and replaces the imp module.
This is important because python3.12 removes the imp module entirely and
any code using imp needs to be rewritten to be python3.12 compatible.

The rewrite here comes from the suggestions in an upstream discussion
about this transition [0].

[0] https://discuss.python.org/t/how-do-i-migrate-from-imp/27885

Change-Id: Iaa742c71bbb208225726c2b8fa5af111f48667c9