Optimize db prune query
We've added some more tables and the query used to delete old
buildsets suffered as a result.  The current approach performs
quite a lot of individual queries: one for each table multiplied
by each buildset that we inspect.

A faster approach is to use one query for each table, regardless
of the number of buildsets (note, however, that sqlalchemy will
shard this in order to manage query length; in practice, this means
one query for each table for every group of 500 buildsets -- so
a 1/500 reduction in the number of queries).

In local testing, this runs in 1/3 of the time as the current
code in postgres, and between 1/5 and 2/3 of the time with
mysql (performance appears more variable than postgres).

Change-Id: Ie93318ab12832516640ee9210ec071e18623eac4