Only consider branches protected when name matches
When a branch is renamed in Github the REST API will also return the new
branch information via the old branche's API endpoint.

This can lead to cases where an old branch is added to the branch cache
as protected even though the branch no longer exists. This is not a
problem for the schedulers, but since there won't be any cached config
in Zookeeper, zuul-web will display a warning about missing config files
for the branch.

Since the REST endpoint also returns the (new) name of the branch we can
validate this against requested branch name.

Change-Id: I2670f951cac1bf41c6569f5495a60e9de262d4a4