Only consider branches protected when name matches
When a branch is renamed in Github the REST API will also return the new
branch information via the old branches API endpoint.

This can lead to cases where an old branch is added as protected to the
branch cache even though the branch longer exists. This is not problem
for the schedulers, but since there won't be any cached config in
Zookeeper, zuul-web will display a warning about missing config files
for the branch.

Since the REST endpoint also returns the name of the branch we can
validate this against requested branch name.

Change-Id: I2670f951cac1bf41c6569f5495a60e9de262d4a4