Avoid joining the provides table in the builds query
The builds query includes an outer join on the provides table for
the purpose of allowing the scheduler to find previously completed
builds which provide artifacts for items currently in the queue.
Otherwise it is unused (it is not possible to query builds by
provides in the web api).

This join can sometimes produce very slow queries, especially
under postgres with certain data characteristics (usually when there
are too few rows to prompt the query planner to use indexes).  This
is similar to the problem that prompted the recent change to the
job runtime queries.

To avoid this, only join the provides table if required for the query.

Change-Id: I83810577c230bcd7365504e01e2cd1d1e642fa0e