gitlab - avoid trigger build when reviewers added/removed
This change fixes an unexpected behavior where Zuul triggers
the buildset when a reviewer is added or removed but also
in case of comment thread resolved.

Now the driver only rely on the 'update' action specifying
the 'oldrev' attribute. Actually this is stated in the doc
that this attribute is set in case of code change.

https://docs.gitlab.com/ee/user/project/integrations/webhook_events.html#merge-request-events

Change-Id: Ibee53e3e9ead9a0bbfdc0d60a35dcdd4b0a0dba7