gitlab doc - update the reference pipelines
This change removes the approval change in the gate pipeline. Indeed it
is not needed to reset the approval in the start and success reporters.

Furthermore and it is the main motivation behind that change is that,
is case of merge failures (for whatever reason) and if gate job succeeded,
then the approval is set and this generates a gate trigger event. This
makes Zuul to re-run the gate pipeline and try to merge again. This
can loop as long as the merge action fails.

Change-Id: I4c22e562797505d8ebdceb75180c537787cc3cb8