Add support for persisting add_host across playbooks
This is intended to address the use case of using a Nodepool namespace
with a job, creating k8s resources within the job, and then addressing
those resources in the normal way for a Zuul job.  The resource creation
can be placed in one playbook while the actual testing can happen in
another (or even several).

In case there are base job playbooks that would behave badly with the
dynamic resources, this also includes a mechanism for the job to revert
the inventory back to the job's original inventory.

Change-Id: Id450f7aeb65ebf824a1c0ef6e9372cf7aea2d310