Add backoff delay only after successful job start
When there is a burst of new jobs, executors could be slowed by the
backoff delay in accepting those jobs. The backoff delay is calculated
based on the number of currently running jobs and is supposed to
distribute jobs more evenly among executors.

So far this delay was added for every processed build request
independent of whether the job could be started or not. With this change
we only add the backoff delay when the previous request could be
handled.

Change-Id: Ic52f3d82ed12fd45922d9ec9d61e8ded43507837