WIP: Use git native commit with squash merge
When squash-merging, a two-part process is used: the commit is
squash-merged into the current index and then the index is committed
to the repo.  The gitpython commit method is slower than the cgit
commit method.  Since we don't need to use the gitpython method, and
it can be cpu and io intensive, have it spawn a cgit process instead.
This offloads that work from the python interpreter (as is done for
the merge operation itself in all the merge methods, including squash).

Change-Id: Ia18a5b86604e8fead7d4b4ed67cd26c5252d1163