Clean up safety check
The safety check was originally written to detect when a dependncy
cycle changed without the pipeline manager noticing.

Since the dependency cycle refactor, items can have multiple changes
and previous processes that were designed around updates to items
causing cascading updates to items behind them (but in the same bundle)
no longer make as much sense.  However, the "safety check" now seems
to make more sense as the primary method for determining that a
dependency cycle has changed.  It fits in well with other checks
in pipeline processing now that it examines the situation for a
single item.

Resolve the temporary safety check by keeping it.  It is cleaned up
a bit and moved earlier in the pipeline processing.  We can also
clean up the slightly awkward silent dequeue/re-enqueue method and
incorporate it into the safety check.  Now the process is:

  If a dpendency cycle changes, dequeue the item without reporting
  and then re-enqueue all of the items changes.

This means that if a dependency cycle (no matter how large) is
split in half, we will keep both halves (now separate) in the
pipeline.  This behavior is likely to be the most intuitive to
users.

In general, there are two ways to update a dependency cycle: with a
new patchset that changes the graph (typically only gerrit) or with
a PR message or topic change (gerrit and others).  To achieve some
consistency between these methods, we reuse the same re-enqueue method
in both cases (but in the case of a patchset superceding a change,
we don't re-enqueue the old change; but we do expect the
patchset-created event to enqueue the new version).  The timing is
still a little different, but the results are the same.

Change-Id: Ifa42b081cbd103ef04d8814c27ab5c51aa5e8335