Use exec in tutorial docker-compose files
In the tutorial docker-compose files, we override the zuul container
commands to run a shell script to wait for certs before starting
Zuul.  When `podman-compose down` runs `podman -t 10 stop`, it sends a
signal to dumb-init which sends a signal to the shell we configured
as well as to the zuul process.  The zuul process starts to shutdown
but the shell kills it before it finishes.

If the process exits before releasing the ssh stream event election,
then the quick-start job runs one of the follow-on tutorials which
restart the scheduler, then the scheduler will wait for the election
timeout of 30 seconds before it starts receiving events.  If the
tutorial playbook has proceeded to emit an event (eg "recheck"), then
the event will be missed and the job fail.

To correct this, exec the zuul processes after our shell script is
done so that only zuul receives the kill signal and handles it
completely.  Same for zk on general principle.

Change-Id: I90edbc68712bbd92ad1ac98b5960347de29b5342