Remove updateJobParentData method
This method was added as part of the initial deduplication work in
959a0b9834cc4c757eec74fabed7792840532b49.  Since we now collect
parent data at the time that we run the job, this method doesn't
actually do anything other than decide when jobs are ready to run.

This change moves that logic back into the findJobsToRun method
and removes the unecessary updateJobParentData method.

Change-Id: Iac744a24ee3902360eeaef371808657a8eeb2080