Increase timeouts in TestComponentRegistry
The test_component_registry test has been flaky recently.  The cause
is not clear, but one potential cause is a test node that is slow
enough that we hit our 10 second timeouts waiting for zk events
or thread scheduling.  To attempt to improve the test, increase the
timeouts we use for state changes.  25 seconds is used to avoid
interation with the 30 second zk connection timeout.

Change-Id: I263f853f57f64252f651c898897536afdb034063