Refresh Keycloak OpenID Howto for latest Keycloak
Going back through the Howto, many of the Keycloak admin UI screens
have changed layout and prompts, making the guide less
straightforward to follow. Update it to match Keycloak 23 (latest at
time of writing).

It also appears that pyJWT 2.5.0 fixed the bug which caused it to
break on encountering unsupported keys, so the noted step for
disabling rsa-enc-generated should no longer be necessary (this has
also become somewhat harder to do in more recent Keycloak versions).

Change-Id: I585439d4718fafd7075a3ed31ebcd10139dc355a