Fix test_timer_preserve_jobs race
This test was racing the assertFinalState method which checks that
all buildsets are complete.  But we might leave timer builds running
at the end of the job.

To avoid this, adopt the model used by most timer tests: add and
remove the timer pipeline configuration during the test so that the
system is idle at shutdown.

Change-Id: I2d9a7761686ddb0263bbac1b9e8b3cbc476c22b1