Test with Keycloak 23
The current Keycloak version is 23, represented by a moving 23.0 tag
on Quay. Run the Zuul integration with this version since it's
probably closer to what organizations are likely to deploy these
days (it's what the OpenDev Collaboratory is currently upgrading
their deployment to).

Drop the database vendor override to H2, since it's the default in
these newer Quarkus-based images anyway (and the envvar which was
used to set this in the older Wildfly-based images is ignored as
best I can tell, so it really was doing nothing).

Change-Id: Ifb0b085bb3702e52db8b382c28e1580643e3c33c