Enable TCP keepalives in zuul_stream
Some remote commands can run for some time without producing output.
Some network devices/firewalls/pnats/etc may decide to drop the
connection since it looks idle.  Try to avoid that by sending TCP
keepalives, and if that doesn't work, that should at least cause
us to be notified that the connection is dead.  In that case, the
log streamer will receive a timeout error on the socket recv call.
We will log it and stop streaming.

Change-Id: Ib8e2aae0de07ff2dcc9de8e244f09e0cf9796e20