Fix a few python type related issues
In zuul/driver/__init__.py and zuul/driver/sql/sqlreporter.py we fix
some method signatures so that they match those of their parents.

In jwt.py we set a default value on the issuer_id attribute as later
processing performed on this value is invalid on None (the default
return value from .get()). Note that if we do get this default value we
should end up trying to do a url request against '/'. When this fails
not instead of only logging:

  AttributeError: 'NoneType' object has no attribute 'endswith'

We should now log:

  requests.exceptions.MissingSchema: Invalid URL '/.well-known/openid-configuration': No scheme supplied. Perhaps you meant https:///.well-known/openid-configuration?

And return this to the user:

  There was an error while fetching OpenID configuration, check logs for details

Finally in zuul/zk/__init__.py we initialize a time counter value to 0.0
instead of 0 as later operations will result in floats. This removes the
explicit type specification of int (which was incorrect).

Change-Id: I76e7295c1d40188db05b85989eb05f8c13ea6405