Implement new status page
The current status page shows a lot of information about the different
pipelines, queues and even the individual queue items and their jobs.
This information can become quite overwhelming - especially for larger
tenants with a lot of pipelines and longer (gate) queues.

For operators, this makes it hard to get a clear picture of what's going
on and for the individual developers it's hard to find their individual
changes (or queues).

The new status page tries to provide a more "operator-oriented" view by
showing the most imporant information about a tenant in a very concised
manner. The idea is to focus only on the relevant information to get an
answer to the following questions:

  1. What is the current state of the tenant?
  2. Which items are failing?
  3. Which pipelines/queues are succeeding (or failing)?
  4. Which queues are piling up items?

To achieve this, the new status page focuses on higher-level information
like pipelines and queues, but doesn't show more detailed ones like the
individual queue items and their jobs. Also, the queues are much more
prominent in this new status view.

To compensate the missing information, the next change implements an
additional "pipeline details view".

The general idea behind this rework is to provide both a
"operator-oriented" view and a "developer-oriented" view without mixing
both into a single view (which is kind of what the current status view
tries to do). Instead of showing all information on a single page, the
new approach splits that information into different scopes:

  1. status / pipeline overview
  2. pipeline and queue details
  3. individual change

Based on those scopes, the pages that handle those focus only on the
relevant information, but don't show too many details about the other
scopes. The changes in this stack currently cover 1. and 2., whereby 3.
is already existing in form of the "single change panel/view".

The change doesn't touch the current StatusPage, but creates a new one
for easier review. The old status page will be cleaned up in a follow-up
commit.

Change-Id: I255e84ba562bb87da6bdbd44b51adbcc73136c47